% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XML_transfer.R
\name{convert_annot_to_geoJSON}
\alias{convert_annot_to_geoJSON}
\title{Convert Annotations to QuPath geoJSON file}
\usage{
convert_annot_to_geoJSON(annotations, outfile, toJSONlist = T, parallel = T)
}
\arguments{
\item{annotations}{HALO-compatible .annotations file either created in HALO or by rSAI}

\item{outfile}{Output file name}

\item{toJSONlist}{File format (list of JSON features \link{default/T} OR FeatureCollection)}

\item{parallel}{Use parallelisation to speed up function (recommended for multi-core processors)}
}
\description{
This function converts a HALO-compatible .annotations file to a QuPath-compatible geoJSON file.
}
