% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XML_transfer.R
\name{convert_geoJSON_to_annot}
\alias{convert_geoJSON_to_annot}
\title{Convert QuPath geoJSON file to Annotations}
\usage{
convert_geoJSON_to_annot(
  geoJSON,
  prefix = "sample_json",
  fromJSONlist = T,
  onelayer = T,
  parallel = T
)
}
\arguments{
\item{fromJSONlist}{File format (list of JSON features \link{default/T} OR FeatureCollection)}

\item{onelayer}{Output both multilayer and 1layer format annotation files; needed for serial-section registration}

\item{parallel}{Use parallelisation to speed up function (recommended for multi-core processors)}

\item{annotations}{QuPath-compatible geoJSON file}

\item{outfile}{Output file name}
}
\description{
This function converts a QuPath-compatible geoJSON file to a HALO-compatible .annotations file.
}
