% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XML_transfer.R
\name{convert_registered_1layer}
\alias{convert_registered_1layer}
\title{Convert 1layer Annotations to Multilayer}
\usage{
convert_registered_1layer(seurat_data, spot_anot, reg_spot_anot, parallel = T)
}
\arguments{
\item{seurat_data}{Seurat data to be annotated}

\item{spot_anot}{Spot polygon annotations file (filepath) with multilayers produced by "export_spot_annotation" family function (rSAI)}

\item{reg_spot_anot}{1layer format spot annotations file (filepath) to convert to multilayer format}

\item{parallel}{Use parallelisation to speed up function (recommended for multi-core processors)}

\item{outfile}{Name of output file}
}
\description{
This function loads a spot annotation file in 1layer format and uses the spot names
from a submitted Seurat data object to convert the spot annotations to multilayer format.
This function is intended for use once annotations have been registered using HALO from
the 1layer spot annotations but can also be used to convert any 1layer format spot
annotations file in to multilayer format with matched spots.
}
