% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XML_transfer.R
\name{export_spot_annotation_fromBase}
\alias{export_spot_annotation_fromBase}
\title{Produce Spot Annotation Polygons from Space Ranger Outputs}
\usage{
export_spot_annotation_fromBase(
  folderpath,
  imagepath,
  prefix = "sample",
  ellipse = F,
  onelayer = T,
  parallel = T
)
}
\arguments{
\item{folderpath}{Folderpath of Space Ranger outputs (path)}

\item{imagepath}{Filepath of full-resolution TIFF image used for Space Ranger input}

\item{prefix}{A prefix used for the output annotation files; can include a directory structure}

\item{ellipse}{Output annotations using ellipse shape format rather than polygon format; faster but not compatible with region functions}

\item{onelayer}{Output both multilayer and 1layer format annotation files; needed for serial-section registration}

\item{parallel}{Use parallelisation to speed up function (recommended for multi-core processors)}
}
\description{
This function provides a method for producing HALO-compatible annotation files
containing polygon vertices (and/or ellipse vertices) for 10x Visium spot data.
The inputs for this function are the Space Ranger outputs along with the
full-resolution TIFF image used as input to Space Ranger. The TIFF image is not
loaded in to memory and is only used for extracting the tagged image resolution
this is necessary as the defined spot radii defined by Space Ranger are inaccurate
leading to spot diameters of 65um (where 55um is the manufacturer definition; current
for Space Ranger version 1.3). Annotations using multilayer format are produced where
each spot resides within a named layer of the annotation file; 1layer format spot
annotations are intended for use when registering coordinates between serial-sections
(the 1layer coordinates are registered using HALO and then reconverted to multilayer
format via "convert_registered_1layer"). This function may take a while to run but
should only need to be ran once per sample.
}
