% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XML_transfer.R
\name{export_spot_annotation_fromSeurat}
\alias{export_spot_annotation_fromSeurat}
\title{Reproduce Spot Annotations using Filtered Seurat Object}
\usage{
export_spot_annotation_fromSeurat(
  seurat_data,
  base_spot_anot,
  image_name = "slice1",
  prefix = "sample_filtered",
  onelayer = T,
  parallel = T
)
}
\arguments{
\item{seurat_data}{Filtered Seurat object to be used for the filtered output}

\item{base_spot_anot}{Base spot annotation file (multilayer) produced by "export_spot_annotation_fromBase"}

\item{image_name}{Name of the image coordinates within the Seurat object to be used (default is "slice1"; \code{seurat_data@images[["slice1"]]})}

\item{prefix}{A prefix used for the output annotation files; can include a directory structure}

\item{onelayer}{Output both multilayer and 1layer format annotation files; needed for serial-section registration}

\item{parallel}{Use parallelisation to speed up function (recommended for multi-core processors)}
}
\description{
This function is intended for use once a base spot annotations file has been produced
by "export_spot_annotation_fromBase" and any filtration of spots has occurred. This
function takes as input the base spot annotations along with the filtered Seurat object
and outputs a filtered multilayer spot annotation file (allowing also the output of
1layer format annotations).
}
