% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_anot.R
\name{import_exclude_region_Visium}
\alias{import_exclude_region_Visium}
\title{Import an Exclusion Annotation}
\usage{
import_exclude_region_Visium(
  seurat_data,
  spot_polygon,
  region_polygon,
  overlap = 70
)
}
\arguments{
\item{seurat_data}{Seurat data to be filtered}

\item{spot_polygon}{Spot polygon annotations file (filepath) with multilayers produced by rSAI}

\item{region_polygon}{Region polygon annotations file (filepath) with multiple layers (if multiple regions) produced by HALO}

\item{overlap}{Percentage overlap required to include a spot within an annotation region, marking the spot for filtration (default=70)}
}
\value{
The submitted Seurat object filtered to remove spots within annotation regions
}
\description{
This function loads a region annotation file produced in HALO and subsets a Seurat object to
include only spots classfied as outside of the supplied annotation region/s based upon a supplied cutoff.
This function requires the spot annotation produced by an "export_spot_annotation" family function.
Spot annotations must match the supplied Seurat object.
}
