% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_anot.R
\name{import_include_region_Visium}
\alias{import_include_region_Visium}
\title{Import an Inclusion Annotation}
\usage{
import_include_region_Visium(
  seurat_data,
  spot_polygon,
  region_polygon,
  overlap = 20
)
}
\arguments{
\item{seurat_data}{Seurat data to be annotated}

\item{spot_polygon}{Spot polygon annotations file (filepath) with multilayers produced by "export_spot_annotation" family function (rSAI)}

\item{region_polygon}{Region polygon annotations file (filepath) with multiple layers (if multiple regions) produced by HALO}

\item{overlap}{Percentage overlap required to include a spot within an annotation region (default=20)}
}
\value{
The submitted Seurat object with annotated regions added to metadata
}
\description{
This function loads a region annotation file produced in HALO and edits a Seurat object metadata,
displaying the overlap of every spot with the region/s and classifies each spot as within or
outside of a region based upon a supplied cutoff. This function requires the spot annotation
produced by an "export_spot_annotation" family function. Spot annotations must match the supplied
Seurat object.
}
