% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XML_transfer.R
\name{split_annotations}
\alias{split_annotations}
\title{Split Annotations File in to Layers}
\usage{
split_annotations(spot_anot, region_df, outfile)
}
\arguments{
\item{spot_anot}{Spot polygon annotations file (filepath) with multilayers produced by "export_spot_annotation" family function (rSAI)}

\item{region_df}{Region data.frame containing two columns: col1 - barcodes/spot names, col2 - allocated region (no region = NA)}

\item{outfile}{Name of output file}
}
\description{
This function takes a table of barcodes with their allocated regions along with a spot
annotation file and produces a filtered spot annotation file containing only spots
within the defined regions. Spots within the same region are stored within 1 layer.
This function is intended for visualisation of regions within HALO.
}
